hiv and bbbp refactor tree cycle fixed
