+++ src/n_dataset/generators/hiv.py
@@ -1,0 +1,10 @@
+from os.path import join
+
+from src.n_dataset.generators.mol_gen import MolGenerator
+
+# dummy class for HIV dataset
+# important parameters
+#     "data_file_name": "HIV.csv"
+#     "data_label_name": "HIV_active"
+class HIV(MolGenerator):
+    pass
